Diseño de Iterator y refactorings
Se ha añadido el patrón iterador.
Se ha movido el paquete de factorygastos a modelos.
Se ha cambiado la propiedad de la clase Gasto:
int tipo a CategoriaGasto tipo
y se ha corregido el resto de código para que funcione correctamente.

+++ src/models/DAO/JDBC/CategoriaGastoDAOJdbc.java
@@ -43,7 +43,7 @@ public class CategoriaGastoDAOJdbc implements CategoriaGastoDAO{
             
             query = this.c.prepareStatement(queryText, Statement.RETURN_GENERATED_KEYS);
             query.setString(1, ct.getName());
-            query.setFloat(2, ct.getIva());
+            query.setDouble(2, ct.getIva());
             query.execute();
             ResultSet rs = query.getGeneratedKeys();
             if (rs.next()) {
